Moved activity classes to a new folder to keep the main folder clean. Began writing comments and documentation for some of the more recently added things such as the DAOProvider and CreatureColiseum.
