+++ app/src/main/java/com/example/project2/database/DAOProvider.java
@@ -2,7 +2,8 @@ package com.example.project2.database;
 /**
  * This will be used to handle DAO's as global singletons
  * because the individual DAO's will not be changing
- * throughout the project
+ * throughout the project. This helps cut down on rewriting
+ * code and keeps things looking cleaner
  */
 
 
@@ -10,10 +11,21 @@ import android.app.Application;
 
 public class DAOProvider {
 
+    //private variable for the UserDAO as a global singleton
     private static UserDAO userDAO;
+
+    //private variable for the AbilityDAO as a global singleton
     private static AbilityDAO abilityDAO;
+
+    //private variable for the CreatureDAO as a global singleton
     private static CreatureDAO creatureDAO;
 
+    /**
+     * This method will be called by the CreatureColiseum class
+     * on start up so that these global instances are guaranteed
+     * to be available when needed.
+     * @param applicationContext
+     */
     public static void init(Application applicationContext) {
         ApplicationDatabase db = ApplicationDatabase.getDatabase(applicationContext);
         userDAO = db.UserDAO();
