+++ app/src/main/res/layout/activity_add_monsters.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddMonstersActivity">
+    tools:context=".activities.AddMonstersActivity">
 
     <TextView
         android:id="@+id/textView2"
