+++ app/src/main/res/layout/activity_admin_landing.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AdminLandingActivity">
+    tools:context=".activities.AdminLandingActivity">
 
     <TextView
         android:id="@+id/textView"
