+++ app/src/main/res/layout/activity_build_creature_detail.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BuildCreatureDetailActivity">
+    tools:context=".activities.BuildCreatureDetailActivity">
 
     <TextView
         android:layout_width="wrap_content"
