+++ app/src/main/res/layout/activity_build_creature_to_add_to_team.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BuildCreatureToAddToTeamActivity">
+    tools:context=".activities.BuildCreatureToAddToTeamActivity">
 
     <Space
         android:id="@+id/userDisplaySpacer"
