+++ app/src/main/res/layout/activity_creature_view_and_editor.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CreatureViewAndEditorActivity">
+    tools:context=".activities.CreatureViewAndEditorActivity">
 
     <Space
         android:id="@+id/userDisplaySpacer"
