+++ app/src/main/res/layout/activity_opponent_select.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".OpponentSelectActivity">
+    tools:context=".activities.OpponentSelectActivity">
 
     <TextView
         android:id="@+id/opponentSelectTextView"
