+++ app/src/main/res/layout/activity_signup.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background_activities"
-    tools:context=".SignupActivity">
+    tools:context=".activities.SignupActivity">
 
     <ImageView
         android:id="@+id/titleSignUpTextView"
