+++ app/src/main/res/layout/activity_team_viewer.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TeamViewerActivity">
+    tools:context=".activities.TeamViewerActivity">
 
     <Space
         android:id="@+id/userDisplaySpacer"
